requirements.txt
bcrypt==4.0.1
cryptography
fastapi
psycopg2-binary
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
pytest
httpx
pytest-asyncio
aiofiles
jinja2
python-dotenv
PyYAML
pydantic-settings
alembic